code/cluster_manager/requirements.txt
python-consul==0.4.1 